Revert "[TRA-18] 파일 구조 설정"
