--- @docs/magic_words.md
@@ -1,3 +1,0 @@
-## You can use the following magic words to link issues
-
-close, closes, closed, closing fix, fixes, fixed, fixing, resolve, resolves, resolved, resolving, complete, completes, completed, completing
